* Level.cs: Change to Level class structure
Change to track the initial value for the chance the player has to
  roll the next
value in their current sequence.

* XElementExtensions.cs: Added extension for future use

Modified to use BubbleBreak namespace. Potential use for reading and
  writing
XML files.

* BubbleBreak.csproj: Project file changes made by IDE

* ui.png:
* ui.png:
* ui.plist:
* ui.plist: Updated spritesheet for game

* Sequences.cs: Fixed sequence spelling for enums

* GameStartLayer.cs: Added reading of new level logic

Reads the chance to roll the next sequence value. Also reads if the
  various sequences are
available for the given level.

* Info.plist: Changed to target iOS 6.0 as lowest version

* Level-alternate.cs: Alternate level class structure

Experimenting with alternate level class structure to try to load XML
  data into a List<t>

* levelinfo.xml: Changes to level setup data for testing

* LevelLayer.cs: Implemented Sequence Generation

Changes to logic and UI design to implement use of different sequences

* Player.cs: Added ability to track unlocked sequences